Mensajes

+++ protected/controllers/UsuariosController.php
@@ -115,7 +115,10 @@ class UsuariosController extends Controller
 		if(Yii::app()->request->isPostRequest)
 		{
 			// we only allow deletion via POST request
-			$this->loadModel($id)->delete();
+			if ($id != 1)
+				$this->loadModel($id)->delete();
+			else
+				
 
 			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
 			if(!isset($_GET['ajax']))
